Move ruletypes from internal to pkg
