+++ internal/controlplane/handlers_evalstatus.go
@@ -24,9 +24,9 @@ import (
 	propSvc "github.com/mindersec/minder/internal/entities/properties/service"
 	"github.com/mindersec/minder/internal/history"
 	ghprop "github.com/mindersec/minder/internal/providers/github/properties"
-	"github.com/mindersec/minder/internal/ruletypes"
 	"github.com/mindersec/minder/internal/util"
 	minderv1 "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
+	"github.com/mindersec/minder/pkg/ruletypes"
 )
 
 const (
