+++ internal/controlplane/handlers_profile.go
@@ -26,9 +26,9 @@ import (
 	"github.com/mindersec/minder/internal/logger"
 	prof "github.com/mindersec/minder/internal/profiles"
 	ghprop "github.com/mindersec/minder/internal/providers/github/properties"
-	"github.com/mindersec/minder/internal/ruletypes"
 	"github.com/mindersec/minder/internal/util"
 	minderv1 "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
+	"github.com/mindersec/minder/pkg/ruletypes"
 )
 
 // CreateProfile creates a profile for a project
