+++ internal/controlplane/handlers_ruletype.go
@@ -24,9 +24,9 @@ import (
 	"github.com/mindersec/minder/internal/db"
 	"github.com/mindersec/minder/internal/engine/engcontext"
 	"github.com/mindersec/minder/internal/logger"
-	"github.com/mindersec/minder/internal/ruletypes"
 	"github.com/mindersec/minder/internal/util"
 	minderv1 "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
+	"github.com/mindersec/minder/pkg/ruletypes"
 )
 
 var (
