+++ internal/controlplane/server.go
@@ -61,10 +61,10 @@ import (
 	"github.com/mindersec/minder/internal/providers/session"
 	reposvc "github.com/mindersec/minder/internal/repositories"
 	"github.com/mindersec/minder/internal/roles"
-	"github.com/mindersec/minder/internal/ruletypes"
 	"github.com/mindersec/minder/internal/util"
 	pb "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
 	serverconfig "github.com/mindersec/minder/pkg/config/server"
+	"github.com/mindersec/minder/pkg/ruletypes"
 )
 
 const metricsPath = "/metrics"
