+++ internal/engine/rtengine/cache.go
@@ -14,9 +14,9 @@ import (
 	"github.com/mindersec/minder/internal/db"
 	"github.com/mindersec/minder/internal/engine/ingestcache"
 	eoptions "github.com/mindersec/minder/internal/engine/options"
-	"github.com/mindersec/minder/internal/ruletypes"
 	rtengine2 "github.com/mindersec/minder/pkg/engine/v1/rtengine"
 	provinfv1 "github.com/mindersec/minder/pkg/providers/v1"
+	"github.com/mindersec/minder/pkg/ruletypes"
 )
 
 // Cache contains a set of RuleTypeEngine instances
