+++ internal/marketplaces/factory.go
@@ -10,10 +10,10 @@ import (
 
 	sub "github.com/mindersec/minder/internal/marketplaces/subscriptions"
 	"github.com/mindersec/minder/internal/profiles"
-	"github.com/mindersec/minder/internal/ruletypes"
 	"github.com/mindersec/minder/pkg/config/server"
 	"github.com/mindersec/minder/pkg/mindpak"
 	src "github.com/mindersec/minder/pkg/mindpak/sources"
+	"github.com/mindersec/minder/pkg/ruletypes"
 )
 
 // NewMarketplaceFromServiceConfig takes the Minder service config and
