+++ internal/marketplaces/subscriptions/service.go
@@ -16,10 +16,10 @@ import (
 
 	"github.com/mindersec/minder/internal/db"
 	profsvc "github.com/mindersec/minder/internal/profiles"
-	"github.com/mindersec/minder/internal/ruletypes"
 	minderv1 "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
 	"github.com/mindersec/minder/pkg/mindpak"
 	"github.com/mindersec/minder/pkg/mindpak/reader"
+	"github.com/mindersec/minder/pkg/ruletypes"
 )
 
 //go:generate go run go.uber.org/mock/mockgen -package mock_$GOPACKAGE -destination=./mock/$GOFILE -source=./$GOFILE
