+++ internal/profiles/validator.go
@@ -17,10 +17,10 @@ import (
 	"k8s.io/apimachinery/pkg/util/sets"
 
 	"github.com/mindersec/minder/internal/db"
-	"github.com/mindersec/minder/internal/ruletypes"
 	"github.com/mindersec/minder/internal/util"
 	minderv1 "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
 	"github.com/mindersec/minder/pkg/engine/selectors"
+	"github.com/mindersec/minder/pkg/ruletypes"
 )
 
 // Validator encapsulates the logic for validating profiles
