+++ pkg/ruletypes/mock/fixtures/service.go
@@ -10,7 +10,7 @@ package fixtures
 import (
 	"errors"
 
-	mockrulesvc "github.com/mindersec/minder/internal/ruletypes/mock"
+	mockrulesvc "github.com/mindersec/minder/pkg/ruletypes/mock"
 	"go.uber.org/mock/gomock"
 )
