+++ pkg/ruletypes/service.go
@@ -300,34 +300,3 @@ func validateRuleUpdate(existingRecord *db.RuleType, newRuleType *pb.RuleType) e
 
 	return nil
 }
-
-// GetDBReleaseStatusFromPBReleasePhase converts a protobuf release phase to a database release status
-func GetDBReleaseStatusFromPBReleasePhase(in pb.RuleTypeReleasePhase) (*db.ReleaseStatus, error) {
-	sev, err := in.InitializedStringValue()
-	if err != nil {
-		return nil, errors.Join(ErrRuleTypeInvalid, err)
-	}
-	var rel db.ReleaseStatus
-
-	if err := rel.Scan(sev); err != nil {
-		// errors from the `Scan` method appear to be caused entirely by bad
-		// input
-		return nil, errors.Join(ErrRuleTypeInvalid, err)
-	}
-
-	return &rel, nil
-}
-
-// GetPBReleasePhaseFromDBReleaseStatus converts a database release status to a protobuf release phase
-func GetPBReleasePhaseFromDBReleaseStatus(s *db.ReleaseStatus) (pb.RuleTypeReleasePhase, error) {
-	if s == nil {
-		return pb.RuleTypeReleasePhase_RULE_TYPE_RELEASE_PHASE_UNSPECIFIED, nil
-	}
-
-	var rel pb.RuleTypeReleasePhase
-	if err := rel.FromString(string(*s)); err != nil {
-		return pb.RuleTypeReleasePhase_RULE_TYPE_RELEASE_PHASE_UNSPECIFIED, err
-	}
-
-	return rel, nil
-}
