+++ pkg/ruletypes/service_test.go
@@ -17,9 +17,9 @@ import (
 
 	"github.com/mindersec/minder/internal/db"
 	dbf "github.com/mindersec/minder/internal/db/fixtures"
-	"github.com/mindersec/minder/internal/ruletypes"
 	"github.com/mindersec/minder/internal/util/ptr"
 	pb "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
+	"github.com/mindersec/minder/pkg/ruletypes"
 )
 
 // both create and update are bundled together since the testing harness is
